requirements.txt
fastapi==0.115.7
uvicorn==0.22.0
python-dotenv==1.0.1
pydantic==2.10.6
pydantic-settings==2.7.1
SQLAlchemy==2.0.37
# PyMySQL==1.1.1
mysql-connector-python==9.2.0
